refacto: merge two components into one
